renamed package to tokenremove
