+++ src/main/java/elasticsearch/tokenremove/RemoveTokenFilter.java
@@ -1,10 +1,9 @@
-package elasticsearch.remove;
+package elasticsearch.tokenremove;
 
 import org.apache.lucene.analysis.CharArraySet;
 import org.apache.lucene.analysis.TokenFilter;
 import org.apache.lucene.analysis.TokenStream;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
-import org.apache.lucene.analysis.tokenattributes.PositionIncrementAttribute;
 
 import java.io.IOException;
