+++ src/main/resources/plugin-descriptor.properties
@@ -3,4 +3,4 @@ version=${project.version}
 name=${project.artifactId}
 java.version=1.8
 elasticsearch.version=${elasticsearch.version}
-classname=elasticsearch.remove.TokenRemovalPlugin
+classname=elasticsearch.tokenremove.TokenRemovalPlugin
